удален EditProfileView, view перенесен во EditProfileViewController
