+++ FakeNFT/Scenes /Profile/Main/Views/ProfileViewController.swift
@@ -204,15 +204,9 @@ private extension ProfileViewController {
     // MARK: - Actions
     @objc func editBarButtonTapped() {
         print("editBarButton Did Tap")
-        if let presenter = presenter {
-            let editProfileViewController = EditProfileViewController(
-                presenter: presenter
-            )
-            editProfileViewController.modalPresentationStyle = .popover
-            self.present(editProfileViewController, animated: true)
-        } else {
-            print("Presenter is nil")
-        }
+        let editProfileViewController = EditProfileViewController()
+        editProfileViewController.modalPresentationStyle = .popover
+        self.present(editProfileViewController, animated: true)
     }
 }
